pyproject.toml
[project]
name = "fastapi-modular-backend"
version = "0.1.0"
description = "FastAPI backend with modular architecture"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "sqlalchemy==2.0.30",
    "aiosqlite==0.20.0",
    "asyncpg==0.29.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.1.3",
    "python-jose[cryptography]==3.3.0",
    "pydantic-settings==2.2.1",
    "structlog==24.1.0",
    "openai==1.30.1",
    "prometheus-client==0.20.0",
    "httpx==0.27.0",
    "pytest==8.2.2",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
mypy_path = "stubs"

[[tool.mypy.overrides]]
module = ["pydantic_settings"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "fastapi",
    "fastapi.*",
    "pydantic",
    "pydantic.*",
    "sqlalchemy",
    "sqlalchemy.*",
    "passlib",
    "passlib.*",
    "jose",
    "jose.*",
    "structlog",
    "structlog.*",
    "prometheus_client",
    "prometheus_client.*",
]
ignore_missing_imports = true
